Updating requirements step for start_windows.bat
If you are going to use `py` to launch the aplication you should use `py` to update the requirements.  `py` will use the most up to date version of python installed on the machine so if someone has multiple versions of python installed and just uses `pip` it might install the requirments to one version and then they try to launch it with another and it wont work.