Update requirements.txt with tgcrypto
TgCrypto is a high-performance, easy-to-install cryptography library specifically written in C for Pyrogram as a Python extension. It is a replacement for a slower Python-only alternative and implements the cryptographic algorithms Telegram requires, namely: AES-256-IGE, AES-256-CTR and AES-256-CBC.

Pyrogram will automatically make use of TgCrypto when detected, all you need to do is to install it.